Add Triangle, Plane, Quad class.

+++ src/code/include/Sphere.cpp
@@ -5,7 +5,7 @@
 //--------------------------------
 // Constructor implementations
 //--------------------------------
-Sphere::Sphere(const Point3& c_pos, double r, const Color& col, std::shared_ptr<Material> mat):
+Sphere::Sphere(const Point3& c_pos, float r, const Color& col, std::shared_ptr<Material> mat):
 center_position(c_pos), radius(r), color(col), mat(mat){}
 
 //--------------------------------
@@ -14,7 +14,7 @@ center_position(c_pos), radius(r), color(col), mat(mat){}
 Point3 Sphere::getCenter_pos() const{
   return center_position;
 }
-double Sphere::getRadius() const{
+float Sphere::getRadius() const{
     return radius;
 }
 Color Sphere::getColor() const{
@@ -22,8 +22,8 @@ Color Sphere::getColor() const{
 }
 
 bool Sphere::hitRay(const Ray& ray, const Interval& ivl, Hit_record& rec) const {
-  double D; // Discriminant
-  double b, a, c; // Temporary variables for discriminant
+  float D; // Discriminant
+  float b, a, c; // Temporary variables for discriminant
   Vec3 rayDir = ray.getDir();
   Vec3 co = ray.getOrig() - center_position;
 
@@ -38,8 +38,8 @@ bool Sphere::hitRay(const Ray& ray, const Interval& ivl, Hit_record& rec) const 
       return false;
   }
 
-  double D_sqrt = std::sqrt(D);
-  double t = (-b - D_sqrt) / a;
+  float D_sqrt = std::sqrt(D);
+  float t = (-b - D_sqrt) / a;
   if (!ivl.contains(t)) {
       // If the smaller root is outside the interval, try the larger one
       t = (-b + D_sqrt) / a;
